decompl/requirements.txt
fvcore==0.1.5.post20220512
numpy==1.23.1
Pillow==9.3.0
scikit_image==0.19.3
thop==0.1.1.post2207130030
torch==1.13.0
torchvision==0.14.0